Moved all code to a single file
