+++ v2/googlecloud-to-googlecloud/pom.xml
@@ -116,6 +116,11 @@
       <version>2.17.1-change-streams-temp1-SNAPSHOT</version>
     </dependency>
     <dependency>
+      <groupId>com.google.api.grpc</groupId>
+      <artifactId>grpc-google-cloud-bigtable-v2</artifactId>
+      <version>2.17.1-change-streams-temp1-SNAPSHOT</version>
+    </dependency>
+    <dependency>
       <groupId>org.apache.beam</groupId>
       <artifactId>beam-sdks-java-io-hadoop-common</artifactId>
     </dependency>
